Upgrade dependencies
We had a warning that we were using an old version of lodash
that has a security vulnerability. This is almost certainly
not a serious problem, since the packages that use lodash
(eslint and lint-staged) are only used during development,
but out of an abundance of caution I have upgraded all our
dependencies to their latest versions.

(I made an exception for D3, because D3 v5 uses ES6 code,
so would limit the cross-browser compatibility of this
component.)

The new version of rollup uses different configuration options,
so the rollup config files had to be updated too.